server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.4
starlette==0.38.4